+++ R/tuning_spaces_rbv2.R
@@ -10,13 +10,11 @@
 #'
 #' @aliases
 #' mlr_tuning_spaces_classif.glmnet.rbv2
-#' mlr_tuning_spaces_classif.kknn.rbv2
 #' mlr_tuning_spaces_classif.ranger.rbv2
 #' mlr_tuning_spaces_classif.rpart.rbv2
 #' mlr_tuning_spaces_classif.svm.rbv2
 #' mlr_tuning_spaces_classif.xgboost.rbv2
 #' mlr_tuning_spaces_regr.glmnet.rbv2
-#' mlr_tuning_spaces_regr.kknn.rbv2
 #' mlr_tuning_spaces_regr.ranger.rbv2
 #' mlr_tuning_spaces_regr.rpart.rbv2
 #' mlr_tuning_spaces_regr.svm.rbv2
@@ -24,9 +22,6 @@
 #'
 #' @section glmnet tuning space:
 #' `r rd_info(lts("classif.glmnet.rbv2"))`
-#'
-#' @section kknn tuning space:
-#' `r rd_info(lts("classif.kknn.rbv2"))`
 #'
 #' @section ranger tuning space:
 #' `r rd_info(lts("classif.ranger.rbv2"))`
@@ -67,29 +62,6 @@ add_tuning_space(
   learner = "regr.glmnet",
   package = "mlr3learners",
   label = "Regression GLM with RandomBot"
-)
-
-# kknn
-vals = list(
-  k = to_tune(1, 30)
-)
-
-add_tuning_space(
-  id = "classif.kknn.rbv2",
-  values = vals,
-  tags = c("rbv2", "classification"),
-  learner = "classif.kknn",
-  package = "mlr3learners",
-  label = "Classification KKNN with RandomBot"
-)
-
-add_tuning_space(
-  id = "regr.kknn.rbv2",
-  values = vals,
-  tags = c("rbv2", "regression"),
-  learner = "regr.kknn",
-  package = "mlr3learners",
-  label = "Regression KKNN with RandomBot"
 )
 
 # ranger